requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
bcrypt==3.2.0
cffi==1.15.0
click==8.1.2
colorama==0.4.4
fastapi==0.75.1
h11==0.13.0
idna==3.3
iso8601==1.0.2
mypy==0.942
mypy-extensions==0.4.3
pycparser==2.21
pydantic==1.9.0
pypika-tortoise==0.1.4
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
tortoise-orm==0.19.0
typing_extensions==4.1.1
uvicorn==0.17.6
